added the basic folder structure
