+++ src/app/aboutPage/page.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const AboutPage = () => {
+    return (
+        <div className='min-h-screen'>
+            About Page Hare
+        </div>
+    );
+};
+
+export default AboutPage;
