+++ src/app/contactPage/page.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ContactPage = () => {
+    return (
+        <div className='min-h-screen'>
+            Contact Page Hare
+        </div>
+    );
+};
+
+export default ContactPage;
