+++ src/app/myResume/page.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const MyResume = () => {
+    return (
+        <div className='min-h-screen'>
+            <h2>My Resume Hare</h2>
+        </div>
+    );
+};
+
+export default MyResume;
