+++ src/app/resumeTemplates/page.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ResumeTemplates = () => {
+    return (
+        <div className='min-h-screen'>
+            Templates Hare
+        </div>
+    );
+};
+
+export default ResumeTemplates;
